+++ src/prescrypt/front/ast/converter.py
@@ -34,6 +34,9 @@ def convert(node: ast.AST) -> my_ast.AST:
                 children.append(kwargs[k])
             case int() | float() | str() | bool() | bytes() | None:
                 kwargs[k] = v
+            case _ if v is ...:
+                # Ellipsis literal
+                kwargs[k] = v
             case _:  # pragma: no cover
                 msg = f"Unknown type: {type(v)}"
                 raise ValueError(msg)